[MMTC-11] Automated sandbox creation
These changes add a sandbox creation feature to MMTC, so that any user allowed to invoke the `mmtc` program can easily create a fully-independent copy of that MMTC installation (including its output product state) elsewhere on the filesystem.  These sandboxes can be run and manipulated in any fashion without affecting the original MMTC installation.  Additionally, the TelemetrySource interface is enhanced with a new method that is called as a hook during sandbox creation so telemetry source plugins can perform any copying/manipulation of their configuration as required for the new to sandbox to work identically to its source MMTC installation.

This new feature can be invoked with `bin/mmtc create-sandbox <path to new sandbox>`.

Additionally:
- Added the concept of MMTC CLI 'commands'
	- This essentially carries forward the preexisting flow of the 'rollback' command having non-correlation behavior, and is similar to other CLI interfaces that have many individual commands that can be invoked, e.g. podman
	- Supported commands are: [correlation|rollback|create-sandbox].  If no command is specified, 'correlation' is assumed, which keeps existing scripts/automation stable with prior releases
	- Changed main class from TimeCorrelationApp to MmtcCli
	- Separated configuration into an abstract MmtcConfig class and its extending concrete classes (TimeCorrelationAppConfig, RollbackConfig, MmtcSandboxCreatorConfig)
		- This was done primarily to avoid an even larger changeset (because the CLI of each command is different, and the current design ties the CLI definition to the overall configuration)
		- The separation of concerns here is only partially complete, and I expect that we'll move more configuration/logic from MmtcConfig into its subclasses over time
- Changed all URIs in config to filepaths