[#48] Add migration feature to migrate output products forward from MMTC 1.5.1 onward
MMTC output products changed significantly since 1.5.1.  Specifically, the Run History File and Time History File both had changes that would take some time and verification for our users to do manually.

Therefore, as part of MMTC 1.6.0 maintenance, we're added a simple migration capability to check the version of the built-in output products and migrat them.  Output plugins are not yet included in this functionality.

For MMTC commands that are sensitive to the structure of the built-in output products, their versions are now checked before proceeding.  If they are found not to be up to date, MMTC will halt and ask users to manually run the 'migrate' command.

When users run `mmtc migrate`, they will be prompted to confirm, along with specifying a path where MMTC will write a zip file of the current output products for backup.  After migration is complete, MMTC will exit, at which time the user (or automation) can run MMTC as normal.

Closes #48.